Framework elements
The way to use the files is in the coments of the "styles-bonzail.css" and at the top of the "bonzail-framework.html" file.
The PSD file tempt to guide graphic designers with the visuals dimentions to use with Bonzail but i'm not sure if it is clear at this moment. I will probably make a new one.